requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP & Scraping
requests==2.31.0
httpx==0.25.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dateutil==2.8.2
PyYAML==6.0.1
